[ci] [gh-actions] Run the platform job using Github Actions
We move the CI job creating the platform windows installer from
Gitlab-driven Inria custom runners to standard Github actions one.

While we are at it, we significantly cleanup the build script, trying
to avoid as much hackery as possible, and splitting it in logical
stages.

Notwithstanding, the current system is still not very satisfactory, in
particular:

- build time is very long, we should cache cygwin folder at least, and
  only rebuild if the coq-platform ref we use has changed; fortunately
  GH actions do support this
- verbosity of the script is out of control
- we can still cleanup way more things

Please review the list of chosen bugs below, and speak if you think it
is too wide.

closes #6807 , closes #7428 , closes #8046 , closes #8622 , closes ##9401 , closes #11073

Note: this PR was originally started from the nice setup of Jason
Gross [1] for windows CI setup, but it was updated to actually use the
platform after #13598

[1] https://github.com/mit-plv/fiat-crypto/blob/master/.github/workflows/coq-windows.yml[ci] [gh-actions] Initial setup of Windows Github Build