[declare] Improve logical code order
Now that the interface has mostly stabilized, we move code around to
respect internal dependency order.

This will allow us to start sharing more code in the 4 principal
cases, and also paves the way for the full merging of obligations and
the removal of the Proof_ending type in favor of stronger type
abstraction.