[test-suite] async-proofs off in tests with Fail Timeout
Apparently the `Timeout` exception is raised by a signal handled, and
that can happen when the running thread is a worker manager, rather than
the main thread (the one that should be interrupted).

Given that the point of these tests is to test *auto and not the STM,
we disable async proofs forcibly.

(cherry picked from commit 397fb9dfbe09b85d8cbeed5854134a6491372ae4)