Change implementation of Const/Rel/Var annotations.
Instead of optional lists of annotations, they are now a custom datatype
that can be one of the following:
* Assum: For assumptions, or opaque definitions;
* Bare of int: For bare terms, containing a length
* Limit of int: For limit terms, containing a length
* Sized of SVar.t * int: For sized terms, containing a length, and also
  a size variable from the old vector of sizes would have started.
In [annotate_fresh], (co)inductives are assigned fresh size expressions
starting with the given size variable, while C/R/V are assigned a Sized
with the given size variable and the correct length.

This is work towards solving issue #10. However, we cannot yet start
treating all constants as nonopaque, because [collect_annots] will
expand C/R/V annotations into full lists of annotations, which can be
very, very large, and causes the original problem we had before.
The solution is to use a representation for SVars.t that stores size
variables by intervals rather than explicitly as sets (maybe an interval
tree or a segment tree).