[search] Don't bail on non existing kind registration.
This is an alternative to #12537 which is closer to behavior in
previous versions and may be safer for 8.12 than changing the kernel
semantics on module includes.

The updated search in #8855 introduced a way to filter by "command
kinds", but, as explained by Hugo Herbelin, these kinds do not always
exist, so a `Not_found` exception was breaking `Search`.

We now handle the exception properly, no kind-free queries should work
as in 8.11.

Fixes #12525 #12647