Faster implementation of Termops.replace_term.
Instead of lifting the pattern repeatedly at every step, we cache the result
inside a table indexed by the number of liftings. This is faster on branching
terms since we perform less computations, but it is not totally satisfactory
either. Ideally we should instead have an equality test parameterized by lifts,
but this is really hard to do from the API we have.