Updating doc wrt addition of grammar subentry "name" and deprecation of "ident".
For constr, this means clarifying that "ident" is deprecated and to be
replaced by "name". Here, some cleaning shall have to be done at the
end of deprecation phase, when "ident" will take its literal meaning.

For custom notations, this is about documenting the effect of "ident"
and "global" which was yet undocumented.

Note that we anticipate an example in constr working with the literal
meaning of "ident" (temporarily silencing the warning).

Co-authored-by: Théo Zimmermann <theo.zimmi@gmail.com>
Co-authored-by: Jim Fehrle <jim.fehrle@gmail.com>