Avoid producing nake pointers.
The only safe way to distinguish accumulators from other function
is to record the set of all accumulator produced. To avoid leaking
this is recorded in a weak table. Sadly the Weak module does not
allow to use the physical equality function to retrieve values since
it is applied on a copy. Ephemeron does not have this problem.