Do not rely on accumulators having tag 0 in native compilation.
Instead, we add a test for accumulator branches. We cannot use Obj.tag because
the latter is inefficient, so we roll up our own test to check that a value
has a closure tag and the right code.

We keep the reserved status of blocks with tag 0 for native arrays. To do this
we preserve the dummy accumulator constructor for all inductive types and replace
all accumulating branches with an assertion failure.