[micromega] Fix bug#12790
zify used to generate many syntactic positivity constraints when translating a goal
from nat to Z. For instance, to state that the product of 2 integers
is positive. Instead, lia performs an interval analysis that is more semantic.

The bug was that the interval analysis was performed after the
elimination of equations.  The current workaround is to perform
interval analysis before and after eliminating equations.

bla

(cherry picked from commit 74932ad2ff431f49370d8bb0f730a588b192b484)