Do not βι-normalize terms without defined metas in meta_reducible_instance.
This function behaved in a non-uniform way before this patch, as it would
return meta-free terms untouched but would perform what would amount to a
βι-normalization over terms that contained metas, none of which was
defined. A comment seems to indicate that we might have βι-normalized
the term in the first case, but it turns out this is expensive. So,
instead we leverage this non-uniformity and extend it to this situation.

This happens a lot e.g. in auto-like tactics, and this patch should speed up
those quite a bit.