Implemented solving subsizing constraints.
After a term has its type inferred, we solve the constraints generated,
producing a size substitution map, which we then apply to the term and
its type before returning the judgement.
This replaces the earlier strategy of erasing all size annotations in
terms and using global annotations in types.

A lot of unused `judge_of_*` functions in Typeops have been removed as
well, since being unused, whether they would require sizes to be erased
is unclear, and they were cluttering up the file anyway.

Additionally, a Typeops.check function has been added to check a bare
term against a bare type, explicitly continuing the size state and
constraint set between them so that the substitution map produced is
consistent between the two. This essentially replaces the task of
`judge_of_cast`.

The RecCheck.solve function requires finding the (weakly) connected
components of the constraint graph. The existing `scc` function in
WeightedDigraph returns rather the /strongly/ connected components, so
we introduce a new function for the WCCs, also stolen from OCamlGraph.
Further details on `solve` appear in the comments in the file.

Note that as side result, definitions that aren't fixpoints can also be
size-preserving. This closes #4.