Make callback opcodes more generic.
This does not make the code any slower, since

  Is_coq_array(accu) && Is_uint63(sp[0])

and

  !Is_accu(accu) && !Is_accu(sp[0])

take the exact same number of tests to pass in the concrete case.

In the accumulator case, it takes one more test to fail, but we do not
care about the performances then.

(cherry picked from commit b6fadbf6f08860609961f6dbad8a036346925265)