Fix #9569 (notations collect the spine binding variables at printing time).
This allows to know which global references whose basename may be
unexpectedly caught need to be qualified.

Note: the alternative strategy, which is sometimes used, of renaming
the binding variables so as to avoid collisions with the basename of a
global reference is somehow less nice.

(cherry picked from commit 094c395ede48f39caf7a48b663abb0e61769369e)