Use fresher names in eqschemes.
The previous implementation appears to be sound when Mangle Names is
off, but it relies on two fragile assumptions, namely that
next_global_ident_away always returns different identifiers when called
with naming hints which are different after stripping all digits from
the end, and that default_id_of_sort (locally defined) never returns
"HC" or "P", or either of those followed by a string of digits.

These changes make both assumptions unnecessary.

As a side note, it seems odd that fresh is ignoring it's env parameter.