Occur-check modulo the equivalence generated by evar-evar unification problems.
We also do some clarification around extract_conv_pbs:
- evar-sensitive Evarutil.whd_evar is moved to evd.ml
- we synchronize explicitly the extraction with the list of last modified evars

There is still an ad hoc conv problem extraction in evar_refiner.ml
which maybe could be replaced by a plain extraction of all modified
evars. But the code was bugged anyway (not able to see that ?y=?x was
a change of ?x), so maybe easy to change.

The change in Notations4.v is an indirect consequence of the more
precise occur-check used for restricting evars.