Tentative fix of #9107 (too weak occur check).
I don't understand well the code, but it looks wrong that the
occur-check is done in the arguments of the later occurrence of an
evar but not in its first occurrence.

The case here is of an instantiation ?T := A _ ?x@{... ?T ...} where
the occur-check was done in the type of ?x but not in the arguments of
?x.

Note that the occur-check in the arguments is too restrictive in
general. For instance, it would be also ok above to drop the
dependency of ?x in ?T. We cannot afford a loop but postponing could
be better.

This indirectly fixes #5694 by preventing a cyclic evar to
occur. However, this is a case where a restriction might be better (?).