Remove the dependency field from compiled libraries.
This data is actually part of the vo summary, so we were duplicating it in the
marshalled data, once in the library segment and once in the summary segment.

Furthermore, there is no trust issue about vo tinkering, since vo import is
already trusted in coqc mode, and a mismatch would be rediscovered by typing
in coqchk.