[build] [native] Don't assume installed native libraries are in custom output path
In #11581 we introduced the `-native-output-dir` option to allow the
build system to redirect the output of the native compiler.

Unfortunately that patch also modified the default loadpath, which is
now buggy if a library with native is installed.

We thus revert the change to the loadpath handling, so for now
additional native build paths have to be passed with `-nI`.

Note that unfortunately in `link_library` we don't know if the
required library is coming from the build dir or from an installed
dir, as this information is generated from `Require` statements in
`Library.get_used_load_paths`. We thus check and give priority to
files in the build location.

As to make the patch backportable I introduced an extra `stat` system
call which should not be problematic as the cache will be hot for the
second call.

An alternative would be actually to modify loadpath compilation in
`call_compiler` so both include paths would be added if `output_dir`
is not the default, however that seems pretty noisy given the large
path set returned by `!get_load_paths`.

(cherry picked from commit f3cc5b6d514fd77887bc1c0cd8aee31e8951ca98)