[coqinit] Respect order of ML includes
This fixes a regression introduced in #11618, where I didn't realize
that the order of ML includes would be important as users may want to
shadow it.

In general I do believe that shadowing is tricky and the build system
should handle it, but for now makes sense to preserver the behavior.

The fix is not very nice, but we cannot afford to tweak the API as
this should be backported to 8.12.1; there is a pending PR refactoring
a bit more the toplevel init that should clean this up.

Fixes #12771

(cherry picked from commit f8a0b47dee445bb71e0824df576f69dd56511257)