Restore the List.Smart.map original implementation.
Commit 56ff0c9 mangled the code claiming to make it tail-rec, but this is not
the case. In addition to make the code convoluted it also over-allocates for
nothing and breaks the write barrier for fun.

We simply rollback to the (slightly cleaned-up) previous code, that was simpler,
likely faster, and as much tail-rec as the one introduced in 56ff0c9.

(cherry picked from commit dcb325fd6ef432cd3717264c1547b94c7da2f396)