metasyntax.ml: splitting interpretation of modifiers into two parts.
The first part now also includes formats (which can be now be
mentioned for a purely-interpretation notation, e.g. in a where
clause).

The second part is about what concerned "Reserved Notation".

Note that beforehand, there was also a preliminary #[...]-based phase
providing the deprecation modifier.

We reorganize the structures accordingly.

In passing, we add a few stronger tests on duplication of modifiers
and locate the errors about notation modifiers.