Configure default value of -native-compiler
This an implementation of point 2 of CEP coq/ceps#48
https://github.com/coq/ceps/pull/48

Option -native-compiler of the configure script now impacts the
default value of the option -native-compiler of coqc. The
-native-compiler option of the configure script is added an ondemand
value, which becomes the default, thus preserving the previous default
behavior.

The stdlib is still precompiled when configuring with -native-compiler
yes. It is not precompiled otherwise.

(cherry picked from commit 1596cdbb2e97f2353236e35fb07be05efe6d1a84)