Intern application arguments in left-to-right order
This makes it so that we have an application `h a b` with both `a` and
`b` unbound, `a` is the one that is reported (parent commit with my current
compiler setup reports `b` first, and the code does not define which
it should be).

Ideally we would report both but that requires more code.

(cherry picked from commit 81a3f81c046697ff011c93929c75c29f4cd39f2f)