Modular printing algorithm for bench/render_results.
The old code was a mess of handwritten ad-hoc code to print the result table
in a fancy way. Instead of hardcoding everything this patch introduces a
generic function to print a table of data. This will allow extension of the
profiling information displayed to the user in an easy way.